cpufreq/qcom-cpufreq-hw: Don't use core count for end-of-table detection
We currently check that both frequency and core count are identical to
identify the end of the LUT table. However, the core count can change
dynamically based on how many CPUs are active and the device
temperature. If we're scanning the table when either of these are
changing, then we can see the same frequency twice with different core
counts. This results in registering a duplicate frequency with OPP,
which prevents the cpufreq policy from being registered.

To fix this, only check for an identical frequency and ignore the
specific core count. No valid table will have the same frequency listed
twice with different core counts.

Dynamic changes to the LUT core counts are only done for mitigation
purposes, and they're only changing the core count of boost
configurations. The updated boost frequencies will still have a
core_count < max_cores. So even though the table may change after we
read it, our list of published boost frequencies is still accurate.

Change-Id: Id88a0e28934df97906447d739925977e6b5779fa
Signed-off-by: Mike Tipton <mdtipton@codeaurora.org>
Signed-off-by: rsuntk <rissu.ntk@gmail.com>